refactor: quizs -> quizzes 로 변경
